Centralise python resource to AppData/Roaming/BHoM

+++ Python_Engine/Compute/Install.cs
@@ -47,22 +47,38 @@ namespace BH.Engine.Python
             if (!force && Query.IsInstalled()) // python seems installed, so exit
                 return;
 
+            if (!Directory.Exists(Query.EmbeddedPythonHome()))
+                Directory.CreateDirectory(Query.EmbeddedPythonHome());
+
             await Task.Run(() =>
             {
-                var assembly = typeof(Compute).Assembly;
-                var appdata = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
-                var zip = Path.Combine(appdata, $"{EMBEDDED_PYTHON}.zip");
-                var resource_name = EMBEDDED_PYTHON;
-                CopyEmbeddedResourceToFile(assembly, resource_name, zip, force);
+                string resourceName = typeof(Compute).Assembly.GetManifestResourceNames().FirstOrDefault(x => x.Contains(EMBEDDED_PYTHON));
+                if (resourceName == null)
+                    throw new FileNotFoundException($"No embedded python zip resource found for {resourceName}");
+
+                // Copy the python embedded zip file to AppData/Roaming/BHoM/
+                string targetFolder = Query.EmbeddedPythonHome();
+                string targetPath = targetFolder + ".zip";
+                CopyEmbeddedResourceToFile(resourceName, targetPath, force);
+
                 try
                 {
-                    ZipFile.ExtractToDirectory(zip, zip.Replace(".zip", ""));
+                    ZipFile.ExtractToDirectory(targetPath, targetPath.Replace(".zip", ""));
 
                     // allow pip on embedded python installation by unflagging python as embedded
                     // see https://github.com/pypa/pip/issues/4207#issuecomment-281236055
                     File.Delete(Path.Combine(Query.EmbeddedPythonHome(), PYTHON_VERSION + "._pth"));
+
                 }
-                catch { }
+                catch
+                {
+                }
+                finally
+                {
+                    // Clean up
+                    if (File.Exists(targetPath))
+                        File.Delete(targetPath);
+                }
             });
         }
